requirements.txt
flask
flask-sqlalchemy
flask-migrate
flask-login
flask-wtf
python-dotenv
Pillow
email_validator
psycopg2-binary

